5054 - заменила id школы с Long на Integer

+++ src/main/java/com/kataacademy/schoolportal/common/services/schooattribute/SchoolService.java
@@ -10,9 +10,9 @@ public interface SchoolService {
 
     public School save(School user);
 
-    public School getById(Long id);
+    public School getById(Integer id);
 
     public School edit(School school);
 
-    public void delete(Long id);
+    public void delete(Integer id);
 }
